Extract RangeValidationTest
