+++ java/core/src/test/java/sleeper/core/range/RangeTest.java
@@ -669,56 +669,4 @@ public class RangeTest {
         assertThat(doesEmptyTo888OverlapPartition).isTrue();
         assertThat(does888ToNullOverlapPartition).isTrue();
     }
-
-    @Test
-    void shouldNotCreateRangeWithMinLongGreaterThanMax() {
-        // Given
-        Field field = new Field("key", new LongType());
-        Object min = 2L;
-        Object max = 1L;
-
-        // When / Then
-        assertThatThrownBy(() -> new Range(field, min, max))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("Range of field key has minimum greater than maximum");
-    }
-
-    @Test
-    void shouldNotCreateRangeWithMinIntGreaterThanMax() {
-        // Given
-        Field field = new Field("key", new IntType());
-        Object min = 2;
-        Object max = 1;
-
-        // When / Then
-        assertThatThrownBy(() -> new Range(field, min, max))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("Range of field key has minimum greater than maximum");
-    }
-
-    @Test
-    void shouldNotCreateRangeWithMinStringGreaterThanMax() {
-        // Given
-        Field field = new Field("key", new StringType());
-        Object min = "b";
-        Object max = "a";
-
-        // When / Then
-        assertThatThrownBy(() -> new Range(field, min, max))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("Range of field key has minimum greater than maximum");
-    }
-
-    @Test
-    void shouldNotCreateRangeWithMinByteArrayGreaterThanMax() {
-        // Given
-        Field field = new Field("key", new ByteArrayType());
-        Object min = new byte[]{2};
-        Object max = new byte[]{1};
-
-        // When / Then
-        assertThatThrownBy(() -> new Range(field, min, max))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("Range of field key has minimum greater than maximum");
-    }
 }
